requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-import-export==3.1.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
PyYAML==6.0
sqlparse==0.4.3
tablib==3.4.0
tzdata==2022.7
xlrd==2.0.1
xlwt==1.3.0
